CORTX-29122: Updated use of CORTX Secrets
 * solution>common>secrets>content no longer required
 * deploy script will generate a value for any secrets>content
       field that does not have a secret value specified.
 * "secrets>name_ref" can be specified instead of "secrets>name".
       If name_ref is specified then deploy script takes the
       specified name as an existing Secret.  (Error if this
       Secret does not exist.)
 * Add a note at the end of deploy about how to get secret values.

+++ k8_cortx_cloud/destroy-cortx-cloud.sh
@@ -321,19 +321,17 @@ function deleteSecrets()
     printf "########################################################\n"
     printf "# Delete Secrets                                       #\n"
     printf "########################################################\n"
-    output=$(./parse_scripts/parse_yaml.sh $solution_yaml "solution.secrets*.name")
-    IFS=';' read -r -a parsed_secret_name_array <<< "$output"
-    for secret_name in "${parsed_secret_name_array[@]}"
-    do
-        secret_name=$(echo $secret_name | cut -f2 -d'>')
+    secret_name=$(./parse_scripts/parse_yaml.sh $solution_yaml "solution.secrets.name")
+    if [[ ! -z "${secret_name}" ]]; then
+        secret_fname=$(echo "${secret_name}" | cut -f2 -d'>')
         kubectl delete secret $secret_name --namespace=$namespace
-    done
 
-    find $(pwd)/cortx-cloud-helm-pkg/cortx-control -name "secret-*" -delete
-    find $(pwd)/cortx-cloud-helm-pkg/cortx-data -name "secret-*" -delete
-    find $(pwd)/cortx-cloud-helm-pkg/cortx-server -name "secret-*" -delete
-    find $(pwd)/cortx-cloud-helm-pkg/cortx-ha -name "secret-*" -delete
-    find $(pwd)/cortx-cloud-helm-pkg/cortx-client -name "secret-*" -delete
+        find $(pwd)/cortx-cloud-helm-pkg/cortx-control -name "secret-*" -delete
+        find $(pwd)/cortx-cloud-helm-pkg/cortx-data -name "secret-*" -delete
+        find $(pwd)/cortx-cloud-helm-pkg/cortx-server -name "secret-*" -delete
+        find $(pwd)/cortx-cloud-helm-pkg/cortx-ha -name "secret-*" -delete
+        find $(pwd)/cortx-cloud-helm-pkg/cortx-client -name "secret-*" -delete
+    fi
 }
 
 function deleteConsul()
